Updated post method and added helpers
added helpers to reuse in PUT, also added daykind in order for the
system to know what day it is for the traveler arrival, stay or
departure. This way it's going to be easier to automate the creations of
packages, when i did this day offset was no longer needed.

the helpers contains code that eralier where in the post method, moving
on to adding the put method and changing comments from swedish to
english